requirements.txt
Faker==24.14.0
fastapi==0.110.2
pre-commit==3.7.1
pydantic==2.7.1
requests==2.32.3
SQLAlchemy==2.0.31
uvicorn==0.29.0
python-dotenv
pydantic-settings
rich
asgi-correlation-id
python-json-logger
httpx
sqlalchemy[asyncio]
aiosqlite